Remove unused custom ScalaEntityWithStaticSerializer
Spring data does not support the scala-generated versions of the Q-classes, only the java ones. The ScalaEntityWithStaticSerializer was an attempt to add a static field QUser.user (of type QUser, like for the java classes) so that spring data would be happy, but it did not seem to work as the generated java code does not in fact get a static member user, but a method user().